Add rudimentary installation instructions to plugin
Describe how `protoc` finds the plugin binary in the path. A simple
installation can be performed by copying the binary to a directory in
the path.

Also fix the example `protoc` invocation to match the Makefile by
adding an explicit `--plugin` argument.

Also add a few more files to .gitignore that are generated during
plugin build/test.