Revert to Apache 2.0 license verbatim [skip ci] (#504)
Replaced the license with the verbatim contents of the canonical license URL:
https://www.apache.org/licenses/LICENSE-2.0.txt

The license seems to be missing random lines, the entire paragraph at the end,
and should not be replacing the template with a specific copyright as this is the
license, not a copyright statement for the project. Each file has a copyright
statement at the top.