Add http2 transport namespace and config (#1880)
Motivation:

The grpc channel will require some config. This can be dervied from
config which we'll make public for the posix and niots transports. As we
need the config to create the channel, we need all the bits that make up
that config and a namespace to hold them in.

Modifications:

- Add an 'HTTP2ClientTransport' namespace and a nested 'Config' namespace
- Add config and defaults

Result:

Config is in place for the grpc channel