Expose listening address on HTTP2 server transport (#1933)
Motivation:

We should expose the listening address on the H2 server transports so that our client can connect to it.

Modifications:

Exposed a `listeningAddress` async property. It will block until we bind or error, return the listening address while it's bound and listening, and throw an error when it's not listening anymore.

Result:

We can now know what address our server transport's listening on for requests.