Avoid throwing from the stream state machine (#1975)
Motivation:

The stream state machine throws on a number of paths rather than
returning an action for the caller to execute. This extra flow control
path can lead to the state not being correctly updated (and left in the
modifying state).

The only valid error the stream state machine may throw is one for being
in an invalid state. It's safe to throw this error as the state won't
change when an invalid state is reached. This can be encoded using typed
throws.

Modifications:

- Use typed throws in the stream state machine, this required pushing
  typed throws down into the compressor
- A few other parts of the state machine needed modifying to add new
  actions to avoid throwing

Result:

The only error the state machine may throw is for an invalid state.