Add support for TLS on H2 NIOTS server transport (#2040)
## Motivation
We currently have a NIOTS server transport implementation in gRPC v2,
but it doesn't support TLS.

## Modifications
This PR adds support for TLS in the NIOTS-backed HTTP/2 implementation
of the server transport for gRPC v2.
It also adds support for ALPN, to validate that the negotiated protocol,
if required, is HTTP2 or `grpc-exp`. If it's not, an error will be
fired/the channel will be closed, since we don't support H1.

## Result
We now support TLS/ALPN when using the NIOTS server transport in gRPC
V2.