Add support for additional TLS options in NIOTS transport (#18)
This PR adds support for a few additional TLS options in the NIOTS
transport, to bring it in line with the NIOPosix implementation. In
particular, it now supports custom trust roots, which means mTLS can be
enabled when using NIOTS, and E2E testing is possible (because we can
use self-signed certificates).

Some code has been refactored to be shared across transport
implementations, and a small bug with how errors were being surfaced has
been fixed.