Use NIO scheduleCallback API for connection handler timers (#28)
## Motivation

NIO 2.75.0 added new APIs for scheduling callbacks, which can be used to
implement timers with fewer allocations than the previous APIs[^0].

## Modifications

Replace the timers used in the client and server connection handlers
with a new implementation that uses these new NIO APIs.

## Result

This should have no functional impact. However it reduces the
allocations in the connection handler. How many allocations it reduces
will depend on the exact scenario, but running the `echo` server under
Instruments and sending 1000 requests suggests around 10%.

[^0]: https://github.com/apple/swift-nio/releases/tag/2.75.0

---------

Co-authored-by: George Barnett <gbarnett@apple.com>