Separate out the security config from the rest of the transport config
Motivation:

Users must explicitly state whether they want TLS or plaintext. This
config for this is currently nested within the rest of the config for
the transport. In many cases this results in users writing
`.defaults(transportSecurity: .tls(.defaults(...)))` where
`.defaults(...)` often includes various TLS config.
This is verbose and a little bit finicky to use.

Modifications:

- Seprate out the transport security from the rest of the config and
  have the client and server transport take a config _and_ a transport
  security config.
- This allows for APIs which compose better while also allowing for the
  transport config to just use a set of defaults.

Result:

Easier to use API.