Propagate the authority pseudoheader (#33)
Motivation:

We should be sending the ":authority" pseudoheader, somehow this was
missed when doing the stream state machine.

The strategy for propagating the authority will be to:
- Use any override value set on the client transport, otherwise
- Use the value provided by a name resolver (if any), otherwise
- Derive a value from the target address.

Modifications:

- Add client config allowing the authority to be overridden
- Add authority to the name resolver interface
- Propagate the authority through to the stream state machine
- Add a percent encoder, as the authority should be percent encoded
- Remove hostname from the client TLS config and use the authority
instead. This is generally a usability win as clients generally won't
have to set this value themselves.

Result:

Authority is set automatically, and can be overridden if necessary.