Delay closing until the next loop tick when gracefully shutting down
Motivation:

The server cancellation test fails every now and then because the client
sees the stream close unexpectedly. This happens when the server is
gracefully shutting down. Writing the final frame with end stream set
triggers the HTTP/2 handler to emit a stream closed event which the
connection management handler picks up and concludes that the connection
is no longer needed (the server is shutting down and now no streams are
open) so closes it.

However because this close happens during the write, it ends up being
ordered before the flush. This means that the frame isn't actually
written to the network before the connection is closed.

Modifications:

- Delay the close until the next event loop tick when it happens during
  graceful shutdown

Result:

Fewer bugs