Make sure we don't write additional DATA frames when flushing once request is finished (#19)
## Motivation
When a client sends a request, to signal its end to the server, it has
to send an empty DATA frame with END_STREAM set.
In the current implementation, the state machine correctly signals the
handler that there are no more messages to be encoded, and so the client
proceeds to write the empty DATA frame with EOS set when flushing.
However, if subsequent flushes are triggered down the pipeline, the
client will keep writing these empty DATA frames, which is a bug.

## Modifications
Make sure that we only write the empty DATA frame once. If we've already
done it, the client can ignore subsequent flushes, as it is done.