Reset go away state (#49)
Motivation:

The pick first load balancer keeps track of whether the current
subchannel is going away. This is done so that if there isn't yet a
subchannel to roll over too it won't be offered up when the GRPCChannel
asks for a subchannel. However, this state wasn't reset in enough places
within the state machine.

Modifications:

- Reset the is-going-away state when updating the current subchannel and
when the current subchannel becomes idle

Result:

Resolves #40