Add debug channel callbacks (#52)
Motivation:

Sometimes it can be useful to modify the channel pipeline or inspect the
channel when debugging. v1 has hooks to allow users to do this which
proved helpful a number of times. v2 should offer similar behaviour.

Modifications:

- Add 'DebugChannelCallbacks' to the client and server config and wire
them up appropriately.
- Add tests. Extract 'TransportKind' from some tests as it was being
recreated a number of times.

Result:

Users can modify the channel pipeline if they need to.

---------

Co-authored-by: Gus Cairo <me@gustavocairo.com>