Ensure port isn't included in SNI hostname (#76)
Motivation:

gRPC derives the authority from various sources and uses this value for
the SNI server hostname. The authority should include non-standard ports
and the SNI hostname must not include the port. In some cases this
wasn't being respected and the port was being used in SNI this results
in handshake failures.

Modifications:

- Have the Connection sanitize the authority so that it's suitable for
SNI.

Result:

- Fewer handshake failures
- Resolves #71