Fix build when using static Linux SDK (#80)
This fixes a build issue when using the static Linux SDK. We were
branching on `#if os(Linux)` to decide whether a value we are using is
an enum or a statically defined integer: Glibc declares it as an enum,
whereas Darwin and Musl define it as an integer. We should make sure we
are importing Glibc and not Musl when we consider it an enum, as
otherwise the build will fail.