Fix bad transition when server sends invalid headers in stream state machine (#88)
While triaging
https://github.com/grpc/grpc-swift-nio-transport/issues/83 we discovered
an issue with the stream state machine, where we would not transition
the server to `closed` when receiving invalid headers. This is not
correct, because at this point the server is in an unrecoverable state
from the client's perspective, and we should ignore any subsequent
messages.

This PR fixes this transition, which was causing an "invalid state"
error (or runtime crash in debug mode). It now drops any further
messages from the server on the floor, instead of throwing/sending
another error message back to the client.

I've also fixed a problem with receiving headers with a 1xx status code:
they should be ignored, but we were transitioning to closed states
instead.