Expose the peerCertificate from mTLS to ServerContext (#97)
Motivation:

When using gRPC, it's useful to pull the peerCertificate from a
successful mTLS connection for authn and/or authz.

Modifications:

These changes make use of the new transportSpecific field in the
ServerContext. After a successful hanshake has completed, it uses the
provided handler to retrieve the verified peer certificate and populate
the Server Context.

There's also a simple happy path test that's been added for validation
purposes.

Temporarily, this commit also references specific updates in
swift-nio-ssl and grpc-swift.

Result:

This commit allows access to the swift Certificate object in an
interceptor for the Posix HTTP2 transport. It also adds a mechanism for
other transports to allow their own transport specific data.

---------

Co-authored-by: George Barnett <gbarnett@apple.com>