Wait for status to be written before closing (#103)
Motivation:

It's possible for the server to drop writes if the HTTP/2 stream's
outbound flow control buffer is full. The main reason for this is that
NIO's async channel has an 'executeThenClose' function within which we
execute the service provided handler for the RPC, after that returns
NIO's async channel closes the stream by doing a 'close(mode: .all)'.

If writes have been buffered in the HTTP/2 handler then the close will
cause the writes to be dropped.

Modifications:

- Wait for the HTTP/2 stream to close as a result of the status being
written in 'executeThenClose' rather than as a result of 'close'.

Result:

- Resolves #96