Enable loading multiple certs from a PEM file
Motivation:

When passing a PEM file with multiple certificates as trust roots only,
the first one would be loaded. This is unintuitive.

Modifications:

Instead of creating a NIOSSLCertificate directly, use a function that
loads multiple certificates from a PEM file. Add a test case that
creates a certificate chain in a temporary directory to test the
expected behavior.

Result:

Multiple certificates can be loaded from a single file.