Load multiple certificates from a single trust root file (#115)
### Motivation

When passing a PEM file with multiple certificates as trust roots, only
the first one is loaded.

### Modifications

Instead of creating a NIOSSLCertificate directly, use a function that
loads multiple certificates from a PEM file. Add a test case that
creates a certificate chain in a temporary directory to verify the
expected behavior.

### Result

Multiple certificates can be loaded from a single PEM file.