Don't wait for close future if the task is cancelled (#128)
Motivation:

If the server task is cancelled while RPCs are running it should
abruptly close the server transport.

This was regressed in
https://github.com/grpc/grpc-swift-nio-transport/commit/4c96199043b5de4b8ed9ec006956b8040a48f5fb
by waiting for streams to be closed. However, if the task is cancelled
end stream won't be written so the stream will never close of its own
accord meaning that shutdown will wedge.

Modifications:

- only wait for the close future if the task hasn't been cancelled

Result:

Resolves https://github.com/grpc/grpc-swift-2/issues/22