Fix continuation misuse in wrapped channel (#136)
Motivation:

The wrapped channel had a code path which failed pending stream
continuations...except that it didn't which in some cases led to
continuations being leaked.

I also noticed a tiny timing window where it was possible for the
regular clients to leak a continuation.

Modifications:

- Actually resume the continuations
- Fix tiny timing window

Result:

Fewer leaking continuations